+++ guillotina/files/manager.py
@@ -195,7 +195,7 @@ class FileManager(object):
                     },
                     headers={"Content-Range": f"bytes */{file.size}"},
                 )
-        except asyncio.CancelledError:  # pragma: no cover
+        except (asyncio.CancelledError, ConnectionResetError):  # pragma: no cover
             logger.info(f"Range cancelled: {range_request} {self.request}")
             # when supporting range headers, the browser will
             # cancel downloads. This is fine.
@@ -230,7 +230,7 @@ class FileManager(object):
                         **kwargs,
                     )
                 await download_resp.write(chunk)
-        except asyncio.CancelledError:  # pragma: no cover
+        except (asyncio.CancelledError, ConnectionResetError):  # pragma: no cover
             logger.info(f"Download cancelled: {self.request}")
             # when supporting range headers, the browser will
             # cancel downloads. This is fine.